+++ src/Core/Domain/ShowcaseCard/CommandHandler/CloseShowcaseCardHandler.php
@@ -30,7 +30,6 @@ namespace PrestaShop\PrestaShop\Core\Domain\ShowcaseCard\CommandHandler;
 use PrestaShop\PrestaShop\Core\Domain\ShowcaseCard\ConfigurationMap;
 use PrestaShop\PrestaShop\Core\ConfigurationInterface;
 use PrestaShop\PrestaShop\Core\Domain\ShowcaseCard\Command\CloseShowcaseCardCommand;
-use PrestaShop\PrestaShop\Core\Domain\ShowcaseCard\Exception\ShowcaseCardException;
 
 /**
  * Saves the showcase card status to keep it closed
@@ -61,8 +60,6 @@ final class CloseShowcaseCardHandler implements CloseShowcaseCardHandlerInterfac
 
     /**
      * @param CloseShowcaseCardCommand $command
-     *
-     * @throws ShowcaseCardException
      */
     public function handle(CloseShowcaseCardCommand $command)
     {
