+++ tests/Integration/Behaviour/Features/Context/Domain/DebugModeFeatureContext.php
@@ -3,40 +3,29 @@
 namespace Tests\Integration\Behaviour\Features\Context\Domain;
 
 use PHPUnit_Framework_Assert;
-use PHPUnit_Framework_AssertionFailedError;
 use PrestaShop\PrestaShop\Adapter\Debug\DebugMode;
 use PrestaShop\PrestaShop\Core\Domain\Configuration\Command\SwitchDebugModeCommand;
-use Symfony\Component\DependencyInjection\Exception\ServiceCircularReferenceException;
-use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 
 class DebugModeFeatureContext extends AbstractDomainFeatureContext
 {
     /**
      * @When I enable debug mode
-     *
-     * @throws ServiceCircularReferenceException
-     * @throws ServiceNotFoundException
      */
-    public function iEnableDebugMode()
+    public function enableDebugMode()
     {
         $this->getCommandBus()->handle(new SwitchDebugModeCommand(true));
     }
 
     /**
      * @When I disable debug mode
-     *
-     * @throws ServiceCircularReferenceException
-     * @throws ServiceNotFoundException
      */
-    public function iDisableDebugMode()
+    public function disableDebugMode()
     {
         $this->getCommandBus()->handle(new SwitchDebugModeCommand(false));
     }
 
     /**
      * @Then debug mode should be enabled in the configuration
-     *
-     * @throws PHPUnit_Framework_AssertionFailedError
      */
     public function debugModeShouldBeEnabledInTheConfiguration()
     {
@@ -48,8 +37,6 @@ class DebugModeFeatureContext extends AbstractDomainFeatureContext
 
     /**
      * @Then debug mode should be disabled in the configuration
-     *
-     * @throws PHPUnit_Framework_AssertionFailedError
      */
     public function debugModeShouldBeDisabledInTheConfiguration()
     {
