+++ tests/Integration/Behaviour/Features/Context/Domain/ThemeMailTemplatesFeatureContext.php
@@ -4,30 +4,23 @@ namespace Tests\Integration\Behaviour\Features\Context\Domain;
 
 use Behat\Gherkin\Node\TableNode;
 use PHPUnit_Framework_Assert;
-use PHPUnit_Framework_AssertionFailedError;
 use PrestaShop\PrestaShop\Core\Domain\MailTemplate\Command\GenerateThemeMailTemplatesCommand;
-use RuntimeException;
-use Symfony\Component\DependencyInjection\Exception\ServiceCircularReferenceException;
-use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
+use Tests\Integration\Behaviour\Features\Context\Util\BehatTableNodeUtils;
 
 class ThemeMailTemplatesFeatureContext extends AbstractDomainFeatureContext
 {
     /**
-     * @When I generate emails with the following properties:
+     * @When I generate emails with the following details:
      *
      * @param TableNode $table
-     *
-     * @throws RuntimeException
-     * @throws ServiceCircularReferenceException
-     * @throws ServiceNotFoundException
      */
-    public function iGenerateEmailsWithTheFollowingProperties(TableNode $table)
+    public function generateEmailsWithTheFollowingDetails(TableNode $table)
     {
-        $data = $this->extractFirstRowFromProperties($table);
+        $data = BehatTableNodeUtils::extractFirstRowFromProperties($table);
 
-        $overwriteTemplates = $data['overwrite_templates'] === '' ? false : $data['overwrite_templates'];
-        $coreMailsFolder = $data['core_mails_folder'] === '' ? '' : $data['core_mails_folder'];
-        $modulesMailFolder = $data['modules_mail_folder'] === '' ? '' : $data['modules_mail_folder'];
+        $overwriteTemplates = $data['overwrite_templates'] !== '' ?: false;
+        $coreMailsFolder = $data['core_mails_folder'] !== '' ?: '';
+        $modulesMailFolder = $data['modules_mail_folder'] !== '' ?: '';
 
         $this->getCommandBus()->handle(
             new GenerateThemeMailTemplatesCommand(
@@ -44,33 +37,10 @@ class ThemeMailTemplatesFeatureContext extends AbstractDomainFeatureContext
      * @Then mails folder with sub folder :subFolder exists
      *
      * @param string $subFolder
-     *
-     * @throws PHPUnit_Framework_AssertionFailedError
      */
-    public function mailsFolderWithSubfolderExists(string $subFolder)
+    public function mailsFolderWithSubFolderExists(string $subFolder)
     {
         $mailsSubFolder = _PS_MAIL_DIR_ . $subFolder;
         PHPUnit_Framework_Assert::assertTrue(is_dir($mailsSubFolder));
-    }
-
-    /**
-     * duplicated ir Orders pull request
-     *
-     * @param TableNode $table
-     *
-     * @return array
-     *
-     * @throws RuntimeException
-     */
-    private function extractFirstRowFromProperties(TableNode $table): array
-    {
-        $hash = $table->getHash();
-        if (count($hash) != 1) {
-            throw new RuntimeException('Properties are invalid');
-        }
-        /** @var array $data */
-        $data = $hash[0];
-
-        return $data;
     }
 }
